fix: use mypy instead of pyright and fix all errors
