+++ .github/workflows/mypy.yaml
@@ -1,0 +1,24 @@
+name: Type Checking
+
+on:
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  mypy:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python
+      uses: actions/setup-python@v4
+      with:
+        python-version: '3.10'
+    - name: Install uv
+      run: pip install uv
+    - name: Create venv
+      run: uv venv
+    - name: Install dependencies
+      run: |
+        uv sync
+    - name: Run Mypy
+      run: uv run mypy .
