+++ tests/sync_client_test.py
@@ -136,15 +136,15 @@ def test_sync_client(service_url: str, schema: str) -> None:
 
     rec = vec.search([1.0, 2.0], filter={"key_1": "val_1", "key_2": "val_2"})
     assert rec[0][SEARCH_RESULT_CONTENTS_IDX] == "the brown fox"
-    assert rec[0]["contents"] == "the brown fox"
+    assert rec[0]["contents"] == "the brown fox"  # type: ignore
     assert rec[0][SEARCH_RESULT_METADATA_IDX] == {
         "key_1": "val_1",
         "key_2": "val_2",
     }
-    assert rec[0]["metadata"] == {"key_1": "val_1", "key_2": "val_2"}
+    assert rec[0]["metadata"] == {"key_1": "val_1", "key_2": "val_2"}  # type: ignore
     assert isinstance(rec[0][SEARCH_RESULT_METADATA_IDX], dict)
     assert rec[0][SEARCH_RESULT_DISTANCE_IDX] == 0.0009438353921149556
-    assert rec[0]["distance"] == 0.0009438353921149556
+    assert rec[0]["distance"] == 0.0009438353921149556  # type: ignore
 
     rec = vec.search([1.0, 2.0], limit=4, predicates=Predicates("key", "==", "val2"))
     assert len(rec) == 1
@@ -218,7 +218,7 @@ def test_sync_client(service_url: str, schema: str) -> None:
         ]
     )
 
-    def search_date(start_date, end_date, expected):
+    def search_date(start_date: datetime | str | None, end_date: datetime | str | None, expected: int) -> None:
         # using uuid_time_filter
         rec = vec.search(
             [1.0, 2.0],
