+++ timescale_vector/client/__init__.py
@@ -1,0 +1,44 @@
+__all__ = [
+    "SEARCH_RESULT_ID_IDX",
+    "SEARCH_RESULT_METADATA_IDX",
+    "SEARCH_RESULT_CONTENTS_IDX",
+    "SEARCH_RESULT_EMBEDDING_IDX",
+    "SEARCH_RESULT_DISTANCE_IDX",
+    "uuid_from_time",
+    "BaseIndex",
+    "IvfflatIndex",
+    "HNSWIndex",
+    "DiskAnnIndex",
+    "QueryParams",
+    "DiskAnnIndexParams",
+    "IvfflatIndexParams",
+    "HNSWIndexParams",
+    "UUIDTimeRange",
+    "Predicates",
+    "QueryBuilder",
+    "Async",
+    "Sync",
+]
+
+from timescale_vector.client.async_client import Async
+from timescale_vector.client.index import (
+    BaseIndex,
+    DiskAnnIndex,
+    DiskAnnIndexParams,
+    HNSWIndex,
+    HNSWIndexParams,
+    IvfflatIndex,
+    IvfflatIndexParams,
+    QueryParams,
+)
+from timescale_vector.client.predicates import Predicates
+from timescale_vector.client.query_builder import QueryBuilder
+from timescale_vector.client.sync_client import Sync
+from timescale_vector.client.utils import uuid_from_time
+from timescale_vector.client.uuid_time_range import UUIDTimeRange
+
+SEARCH_RESULT_ID_IDX = 0
+SEARCH_RESULT_METADATA_IDX = 1
+SEARCH_RESULT_CONTENTS_IDX = 2
+SEARCH_RESULT_EMBEDDING_IDX = 3
+SEARCH_RESULT_DISTANCE_IDX = 4
